add support for yasso_awen_rate_mod parameter

+++ models/basgra/R/run_BASGRA.R
@@ -360,7 +360,7 @@ run_BASGRA <- function(run_met, run_params, site_harvest, site_fertilize, start_
   #days_harvest <- as.integer(days_harvest)
   
   # run  model
-  NPARAMS = as.integer(144)
+  NPARAMS = as.integer(145)
   if (length(run_params) != NPARAMS) {
     PEcAn.logger::logger.severe(sprintf('%i parameters required, %i given', NPARAMS, length(run_params)))
   }
